add backup of primary site and cdn
ported the mozilla backup script for mdn and extended it to backup the cdn'd assets. added some todo steps that need definition, including where/how to export this.

if using this output to run a mirror, its probably easier to mirror two domains than it is to integrate these two sites and rewrite all the URLs